+++ app/src/main/java/com/muasya/clientapp/Database/LocalCartDataSource.kt
@@ -23,7 +23,7 @@ class LocalCartDataSource(private val cartDAO: CartDAO) : CartDataSource {
         return cartDAO.countItemInCart(uid)
     }
 
-    override fun sumPrice(uid: String): Single<Long> {
+    override fun sumPrice(uid: String): Single<Double> {
         return  cartDAO.sumPrice(uid)
     }
